+++ studio/src/app/stores/busy.store.ts
@@ -4,14 +4,12 @@ interface BusyStore {
   busy: boolean | undefined;
   slideReady: boolean;
   deckReady: boolean;
-  docReady: boolean;
 }
 
 const {state, onChange} = createStore<BusyStore>({
   busy: undefined,
   slideReady: false,
-  deckReady: false,
-  docReady: false
+  deckReady: false
 });
 
 export default {state, onChange};