+++ webcomponents/studio/CHANGELOG.md
@@ -1,3 +1,10 @@
+# 0.0.6 (2022-03-13)
+
+### Features
+
+- remove common "styles"
+- new event `docReady`
+
 # 0.0.5 (2022-03-06)
 
 ### Features